[Rename] 중복 되는 파일 컴포넌트 분리
