--- frontend/src/views/PathPage/TransportationColor.js
@@ -1,83 +1,0 @@
-// TransportationColor.js
-
-// 버스 노선 타입에 따른 색상표
-const busRouteColors = {
-  1: '#36C8F5', // 일반 (시내버스)
-  2: '#F5A623', // 좌석
-  3: '#77DD77', // 마을버스
-  4: '#FF4500', // 직행좌석
-  5: '#FFB6C1', // 공항버스
-  6: '#FF4500', // 간선급행 (직행좌석과 동일)
-  10: '#A52A2A', // 외곽
-  11: '#428df5', // 간선 (서울 간선버스, 파랑버스)
-  12: '#70A800', // 지선 (서울 지선버스, 초록버스)
-  13: '#FAA61A', // 순환 (서울 순환버스, 노랑버스)
-  14: '#B22222', // 광역 (서울 광역버스, 빨강버스)
-  15: '#8B00FF', // 급행 (광역 급행)
-  16: '#FFD700', // 관광버스 (노랑색)
-  20: '#228B22', // 농어촌버스
-  22: '#4682B4', // 경기도 시외형버스 (파랑버스)
-  26: '#DC143C' // 급행간선
-}
-
-// 버스 혼잡도 타입에 따른 색상표
-const busCongestionColors = {
-  '-1': '#808080', // 데이터 없음
-  1: '#00FF00', // 여유
-  2: '#FFFF00', // 보통
-  3: '#FFA500', // 혼잡
-  4: '#FF0000' // 매우 혼잡
-}
-
-// 지하철 노선 타입에 따른 색상표
-const subwayColors = {
-  1: '#0052A4', // 수도권 1호선
-  2: '#00A84D', // 수도권 2호선
-  3: '#EF7C1C', // 수도권 3호선
-  4: '#00A5DE', // 수도권 4호선
-  5: '#996CAC', // 수도권 5호선
-  6: '#CD7C2F', // 수도권 6호선
-  7: '#747F00', // 수도권 7호선
-  8: '#E6186C', // 수도권 8호선
-  9: '#BDB092', // 수도권 9호선
-  91: '#D4003B', // GTX-A
-  101: '#0065B3', // 공항철도
-  102: '#FFCD12', // 자기부상철도
-  104: '#77C4A3', // 경의중앙선
-  107: '#D4003B', // 에버라인
-  108: '#0C8E72', // 경춘선
-  109: '#ED1C24', // 신분당선
-  110: '#A17800', // 의정부경전철
-  112: '#003DA5', // 경강선
-  113: '#B7C452', // 우이신설선
-  114: '#81A914', // 서해선
-  115: '#AD8605', // 김포골드라인
-  116: '#F5A200', // 수인분당선
-  117: '#6789CA', // 신림선
-  21: '#7CA8D5', // 인천 1호선
-  22: '#ED8B00', // 인천 2호선
-  31: '#007448', // 대전 1호선
-  41: '#D93F5C', // 대구 1호선
-  42: '#00AA80', // 대구 2호선
-  43: '#FFB100', // 대구 3호선
-  51: '#009088', // 광주 1호선
-  71: '#F06A00', // 부산 1호선
-  72: '#81BF48', // 부산 2호선
-  73: '#BB8C00', // 부산 3호선
-  74: '#217DCB', // 부산 4호선
-  78: '#81A914', // 동해선
-  79: '#8652A1' // 부산-김해경전철
-}
-
-// 함수 내보내기
-export function getBusRouteColor(routeType) {
-  return busRouteColors[routeType] || busRouteColors[1] // 기본값은 일반 버스
-}
-
-export function getBusCongestionColor(congestionType) {
-  return busCongestionColors[congestionType] || busCongestionColors['-1'] // 기본값은 데이터 없음
-}
-
-export function getSubwayColor(lineType) {
-  return subwayColors[lineType] || subwayColors[1] // 기본값은 수도권 1호선
-}
